Make the strategy for getting order IDs selectable.
